Change: Move the tender to be an abstract cargo type
